backend/requirements.txt
Flask==2.0.2
Jinja2==3.0.2
PyYAML==6.0
Werkzeug==2.0.2
elasticsearch==7.15.2
psycopg2.binary>=2.9.2
redis==4.0.2
urllib3==1.26.7
apscheduler>=3.2.0,<4.0.0
python-dateutil>=2.4.2
SQLAlchemy>=1.4.27
flask-cors>=3.0.10
